+++ R/behaviors.R
@@ -229,6 +229,11 @@ auto_adapt_label <- function(
 #' @param style Style specification for the selection box (list).
 #' See \url{https://g6.antv.antgroup.com/en/manual/behavior/brush-select#style}.
 #' @param trigger Shortcut keys for selection (character vector).
+#' @param outputId Manually pass the Shiny output ID. This is useful when the graph
+#' is initialised outside the shiny render function and the ID cannot be automatically
+#' inferred. This allows to set input values from the callback function
+#' with the right namespace and graph ID. You must typically pass `session$ns("graphid")`
+#' to ensure this also works in modules.
 #' @param ... Extra parameters. See \url{https://g6.antv.antgroup.com/en/manual/behavior/brush-select}.
 #'
 #' @return A list with the configuration settings for the brush select behavior.
@@ -266,6 +271,7 @@ brush_select <- function(
   state = c("selected", "active", "inactive", "disabled", "highlight"),
   style = NULL,
   trigger = "shift",
+  outputId = NULL,
   ...
 ) {
   # Validate inputs
@@ -308,17 +314,16 @@ brush_select <- function(
 
   # We can't access the graph instance from within this JS
   # code as it is create oustide the widget factory
-  output_id <- shiny::getCurrentOutputInfo()[["name"]]
+  if (is.null(config$outputId)) {
+    config$outputId <- shiny::getCurrentOutputInfo()[["name"]]
+  }
 
   # Provide default for onSelect for Shiny context
-  if (is.null(config$onSelect) && !is.null(output_id)) {
-    # unfortunately, we can't create selected_node, selected_edge
-    # since state contains all ids without categories. This could
-    # be possible if IDs were prefixed with their type, e.g., node-1, edge-2, etc.
-    # but that's not the case currently.
+  if (is.null(config$onSelect) && !is.null(config$outputId)) {
     config$onSelect <- JS(
       sprintf(
         "(states) => {
+          //const outputId = this.
           const selected = Object.getOwnPropertyNames(states);
           const nodes = selected
             .filter(id => id.startsWith('node-'))
@@ -336,9 +341,9 @@ brush_select <- function(
 
           return states;
         }",
-        output_id,
-        output_id,
-        output_id
+        config$outputId,
+        config$outputId,
+        config$outputId
       )
     )
   }
@@ -523,6 +528,11 @@ collapse_expand <- function(
 #' creation is succesful so that it does not conflict with other drag behaviors.
 #' @param style Style of the newly created edge (list, default: NULL).
 #' @param notify Whether to show a feedback message in the ui.
+#' @param outputId Manually pass the Shiny output ID. This is useful when the graph
+#' is initialised outside the shiny render function and the ID cannot be automatically
+#' inferred. This allows to set input values from the callback function
+#' with the right namespace and graph ID. You must typically pass `session$ns("graphid")`
+#' to ensure this also works in modules.
 #' @param ... Extra parameters. See \url{https://g6.antv.antgroup.com/en/manual/behavior/create-edge}.
 #'
 #' @note \link{create_edge}, \link{drag_element} and \link{drag_element_force} are incompatible by default,
@@ -542,6 +552,7 @@ create_edge <- function(
   onFinish = NULL,
   style = NULL,
   notify = FALSE,
+  outputId = NULL,
   ...
 ) {
   # Validate inputs
@@ -574,10 +585,12 @@ create_edge <- function(
 
   # We can't access the graph instance from within this JS
   # code as it is create oustide the widget factory
-  output_id <- shiny::getCurrentOutputInfo()[["name"]]
+  if (is.null(config$outputId)) {
+    config$outputId <- shiny::getCurrentOutputInfo()[["name"]]
+  }
 
   # Provide default in Shiny context only
-  if (is.null(config$onFinish) && !is.null(output_id)) {
+  if (is.null(config$onFinish) && !is.null(config$outputId)) {
     config$onFinish <- JS(
       sprintf(
         "(edge) => {
@@ -615,7 +628,7 @@ create_edge <- function(
           }
         }",
         as.numeric(notify),
-        output_id
+        config$outputId
       )
     )
   }
@@ -1189,6 +1202,11 @@ hover_activate <- function(
 #' @param state State to switch to when selected (string, default: "selected").
 #' @param style Style of the lasso during selection (list, default: NULL).
 #' @param trigger Press this shortcut key along with mouse click to select (character vector, default: c("shift")).
+#' @param outputId Manually pass the Shiny output ID. This is useful when the graph
+#' is initialised outside the shiny render function and the ID cannot be automatically
+#' inferred. This allows to set input values from the callback function
+#' with the right namespace and graph ID. You must typically pass `session$ns("graphid")`
+#' to ensure this also works in modules.
 #' @param ... Extra parameters. See \url{https://g6.antv.antgroup.com/en/manual/behavior/lasso-select}.
 #'
 #' @return A list with the configuration settings for the lasso-select behavior.
@@ -1223,6 +1241,7 @@ lasso_select <- function(
   state = "selected",
   style = NULL,
   trigger = c("shift"),
+  outputId = NULL,
   ...
 ) {
   # Validate inputs
@@ -1279,9 +1298,11 @@ lasso_select <- function(
   }
 
   # Recover output id in Shiny context
-  output_id <- shiny::getCurrentOutputInfo()[["name"]]
+  if (is.null(config$outputId)) {
+    config$outputId <- shiny::getCurrentOutputInfo()[["name"]]
+  }
 
-  if (is.null(config$onSelect) && !is.null(output_id)) {
+  if (is.null(config$onSelect) && !is.null(config$outputId)) {
     config$onSelect <- JS(
       sprintf(
         "(states) => {
@@ -1302,9 +1323,9 @@ lasso_select <- function(
 
           return states;
         }",
-        output_id,
-        output_id,
-        output_id
+        config$outputId,
+        config$outputId,
+        config$outputId
       )
     )
   }